requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
beautifulsoup4==4.12.3
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
